+++ zcash_client_backend/src/data_api/testing/pool.rs
@@ -508,6 +508,84 @@ pub fn send_max_single_step_proposed_transfer<T: ShieldedPoolTester>(
     );
 }
 
+/// Tests that attempting to send all the spendable funds within the given shielded pool in a
+/// single transaction fail if there are funds that are not yet confirmed.
+///
+/// The test:
+/// - Adds funds to the wallet in a single note.
+/// - Checks that the wallet balances are correct.
+/// - Mine empty blocks
+/// - Add more funds
+/// - Attempts to construct a request to spend the whole balance to an external address in the
+///   same pool.
+/// - catches failure
+/// - verifies the failure is the one expected
+pub fn send_max_proposal_fails_when_unconfirmed_funds_present<T: ShieldedPoolTester>(
+    dsf: impl DataStoreFactory,
+    cache: impl TestCache,
+) {
+    let mut st = TestBuilder::new()
+        .with_data_store_factory(dsf)
+        .with_block_cache(cache)
+        .with_account_from_sapling_activation(BlockHash([0; 32]))
+        .build();
+
+    let account = st.test_account().cloned().unwrap();
+    let dfvk = T::test_account_fvk(&st);
+
+    // Add funds to the wallet in a single note
+    let value = Zatoshis::const_from_u64(60000);
+
+    let (h1, _, _) = st.generate_next_block(&dfvk, AddressType::DefaultExternal, value);
+    
+    st.generate_empty_block();
+    st.generate_empty_block();
+    let later_on_value = Zatoshis::const_from_u64(123456);
+    let (h, _, _) = st.generate_next_block(&dfvk, AddressType::DefaultExternal, later_on_value);
+    st.scan_cached_blocks(h1, 4);
+
+    assert_eq!(
+        st.wallet()
+            .block_max_scanned()
+            .unwrap()
+            .unwrap()
+            .block_height(),
+        h
+    );
+
+    // Spendable balance doesn't match total balance
+    let total_balance = st.get_total_balance(account.id());
+    let spendable_balance = st.get_spendable_balance(account.id(), 2);
+    assert_ne!(total_balance, spendable_balance);
+
+    let to_extsk = T::sk(&[0xf5; 32]);
+    let to: Address = T::sk_default_address(&to_extsk);
+
+    let fee_rule = StandardFeeRule::Zip317;
+
+    let change_memo = "Test change memo".parse::<Memo>().unwrap();
+    let send_max_memo = "Test Send Max memo".parse::<Memo>().unwrap();
+    let change_strategy = standard::SingleOutputChangeStrategy::new(
+        fee_rule,
+        Some(change_memo.clone().into()),
+        T::SHIELDED_PROTOCOL,
+        DustOutputPolicy::default(),
+    );
+    let input_selector = GreedyInputSelector::new();
+
+    assert_matches!(
+        st
+        .propose_send_max_transfer(
+            account.id(),
+            &input_selector,
+            &change_strategy,
+            to,
+            Some(MemoBytes::from(send_max_memo.clone())),
+            NonZeroU32::new(2).unwrap(),
+        ),
+        Err(data_api::error::Error::DataSource(_))
+    );
+}
 /// This test attempts to send the max spendable funds to a TEX address recipient
 /// checks that the transactions were stred and that the amounts involved are correct
 #[cfg(feature = "transparent-inputs")]