+++ zcash_client_sqlite/src/wallet/common.rs
@@ -188,6 +188,82 @@ where
         return Ok(vec![]);
     }
 
+    let mut stmt_ineligible_notes = conn.prepare_cached(
+        &format!(
+            "WITH ineligible AS (
+                 SELECT
+                     {table_prefix}_received_notes.id AS id, txid, {index_col},
+                     diversifier, value, {note_reconstruction_cols}, commitment_tree_position,
+                     accounts.ufvk as ufvk, recipient_key_scope
+                 FROM {table_prefix}_received_notes
+                 INNER JOIN accounts
+                    ON accounts.id = {table_prefix}_received_notes.account_id
+                 INNER JOIN transactions
+                    ON transactions.id_tx = {table_prefix}_received_notes.tx
+                 WHERE accounts.uuid = :account_uuid
+                 AND {table_prefix}_received_notes.account_id = accounts.id
+                 AND value > 5000 -- FIXME #1316, allow selection of dust inputs
+                 AND accounts.ufvk IS NOT NULL
+                 AND recipient_key_scope IS NOT NULL
+                 AND nf IS NOT NULL
+                 AND commitment_tree_position IS NOT NULL
+                 AND transactions.block > :anchor_height
+                 AND {table_prefix}_received_notes.id NOT IN rarray(:exclude)
+                 AND {table_prefix}_received_notes.id NOT IN (
+                   SELECT {table_prefix}_received_note_id
+                   FROM {table_prefix}_received_note_spends
+                   JOIN transactions stx ON stx.id_tx = transaction_id
+                   WHERE stx.block IS NOT NULL -- the spending tx is mined
+                   OR stx.expiry_height IS NULL -- the spending tx will not expire
+                   OR stx.expiry_height > :anchor_height -- the spending tx is unexpired
+                 )
+                 AND NOT EXISTS (
+                    SELECT 1 FROM v_{table_prefix}_shard_unscanned_ranges unscanned
+                    -- select all the unscanned ranges involving the shard containing this note
+                    WHERE {table_prefix}_received_notes.commitment_tree_position >= unscanned.start_position
+                    AND {table_prefix}_received_notes.commitment_tree_position < unscanned.end_position_exclusive
+                    -- exclude unscanned ranges that start above the anchor height (they don't affect spendability)
+                    AND unscanned.block_range_start <= :anchor_height
+                    -- exclude unscanned ranges that end below the wallet birthday
+                    AND unscanned.block_range_end > :wallet_birthday
+                 )
+             )
+             SELECT 1 
+             FROM ineligible",
+        )
+    )?;
+
+
+    let excluded: Vec<Value> = exclude
+        .iter()
+        .filter_map(|ReceivedNoteId(p, n)| {
+            if *p == protocol {
+                Some(Value::from(*n))
+            } else {
+                None
+            }
+        })
+        .collect();
+    
+    let excluded_ptr = Rc::new(excluded);
+
+    let ineligible_notes = stmt_ineligible_notes.query_and_then(
+        named_params![
+            ":account_uuid": account.0,
+            ":anchor_height": &u32::from(anchor_height),
+            ":exclude": &excluded_ptr,
+            ":wallet_birthday": u32::from(birthday_height)
+        ],
+        |r| to_spendable_note(params, r),
+    )?;
+
+    let unspendable_notes = ineligible_notes
+        .filter_map(|r| r.transpose())
+        .collect::<Result<Vec<ReceivedNote<ReceivedNoteId, Note>>, SqliteClientError>>()?;
+
+    if !unspendable_notes.is_empty() {
+        return Err(SqliteClientError::IneligibleNotes)
+    }
     // The goal of this SQL statement is to select all the spendable notes above the
     // MARGINAL_FEE (5000 zats).
     // 1)  Create a view of all notes, ordered from oldest to
@@ -241,17 +317,6 @@ where
         )
     )?;
 
-    let excluded: Vec<Value> = exclude
-        .iter()
-        .filter_map(|ReceivedNoteId(p, n)| {
-            if *p == protocol {
-                Some(Value::from(*n))
-            } else {
-                None
-            }
-        })
-        .collect();
-    let excluded_ptr = Rc::new(excluded);
 
     let notes = stmt_select_notes.query_and_then(
         named_params![